Change "Ready to send" to "Received"
OK so this is probably my first official pull request on an actual working repository, so please forgive if I'm doing this wrong. 

I thought that "Received" would be a better choice for payroll status that "Ready to send," because received covers the fact that you've got it, whether or not you've reviewed it yet. If you don't want to accept this change (or merge or whatever) no big deal.